Move find_path and find_path_shortest to PAutomaton
and use the null-value of AutomatonPath
