+++ test/Verification_test.cpp
@@ -566,9 +566,9 @@ BOOST_AUTO_TEST_CASE(Verification_longest_trace_arithmetic_3_5_test)
     auto [trace, weight] = Solver::get_trace<Trace_Type::Longest>(instance);
     BOOST_CHECK_EQUAL(w, weight);
 
-    BOOST_CHECK(automaton_path.has_value());
+    BOOST_CHECK(!automaton_path.is_null());
     // TODO: Test instead of printing
-    details::TraceBack tb(instance.automaton(), std::move(automaton_path).value());
+    details::TraceBack tb(instance.automaton(), std::move(automaton_path));
     std::unordered_set<std::tuple<size_t,uint32_t,size_t>, absl::Hash<std::tuple<size_t,uint32_t,size_t>>> seen_front;
     while (seen_front.emplace(tb.path().front_edge()).second) {
         print_conf_path(s, tb.path(), pda);
@@ -654,9 +654,9 @@ BOOST_AUTO_TEST_CASE(Verification_longest_trace_arithmetic_3_3or5_test)
     auto [trace, weight] = Solver::get_trace<Trace_Type::Longest>(instance);
     BOOST_CHECK_EQUAL(w, weight);
 
-    BOOST_CHECK(automaton_path.has_value());
+    BOOST_CHECK(!automaton_path.is_null());
     // TODO: Test instead of printing
-    details::TraceBack tb(instance.automaton(), std::move(automaton_path).value());
+    details::TraceBack tb(instance.automaton(), std::move(automaton_path));
     std::unordered_set<std::tuple<size_t,uint32_t,size_t>, absl::Hash<std::tuple<size_t,uint32_t,size_t>>> seen_front;
     while (seen_front.emplace(tb.path().front_edge()).second) {
         print_conf_path(s, tb.path(), pda);
@@ -714,9 +714,9 @@ BOOST_AUTO_TEST_CASE(Verification_longest_trace_which_pop_seq_test)
     auto [trace, weight] = Solver::get_trace<Trace_Type::Longest>(instance);
     BOOST_CHECK_EQUAL(w, weight);
 
-    BOOST_CHECK(automaton_path.has_value());
+    BOOST_CHECK(!automaton_path.is_null());
     // TODO: Test instead of printing
-    details::TraceBack tb(instance.automaton(), std::move(automaton_path).value());
+    details::TraceBack tb(instance.automaton(), std::move(automaton_path));
     std::unordered_set<std::tuple<size_t,uint32_t,size_t>, absl::Hash<std::tuple<size_t,uint32_t,size_t>>> seen_front;
     while (seen_front.emplace(tb.path().front_edge()).second) {
         print_conf_path(s, tb.path(), pda);
@@ -797,9 +797,9 @@ BOOST_AUTO_TEST_CASE(Verification_longest_trace_hill_test)
     BOOST_CHECK_NE(pXq, nullptr);
     BOOST_CHECK_EQUAL(pXq->second, max_weight<uint32_t>::bottom());
 
-    BOOST_CHECK(automaton_path.has_value());
+    BOOST_CHECK(!automaton_path.is_null());
     // TODO: Test instead of printing
-    details::TraceBack tb(instance.automaton(), std::move(automaton_path).value());
+    details::TraceBack tb(instance.automaton(), std::move(automaton_path));
     std::unordered_set<std::tuple<size_t,uint32_t,size_t>, absl::Hash<std::tuple<size_t,uint32_t,size_t>>> seen_front;
     while (seen_front.emplace(tb.path().front_edge()).second) {
         print_conf_path(s, tb.path(), pda);
@@ -848,9 +848,9 @@ BOOST_AUTO_TEST_CASE(Verification_longest_trace_start_hill_end_test)
     auto [automaton_path, w] = instance.find_path<Trace_Type::Longest>();
     BOOST_CHECK_EQUAL(w, max_weight<uint32_t>::bottom());
 
-    BOOST_CHECK(automaton_path.has_value());
+    BOOST_CHECK(!automaton_path.is_null());
     // TODO: Test instead of printing
-    details::TraceBack tb(instance.automaton(), std::move(automaton_path).value());
+    details::TraceBack tb(instance.automaton(), std::move(automaton_path));
     std::unordered_set<std::tuple<size_t,uint32_t,size_t>, absl::Hash<std::tuple<size_t,uint32_t,size_t>>> seen_front;
     while (seen_front.emplace(tb.path().front_edge()).second) {
         print_conf_path(s, tb.path(), pda);
